requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.1
aiosqlite>=0.19.0
typing-extensions>=4.8.0
asyncio>=3.4.3 